# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "pxr-block-proxy-service",
        "version": "1.0",
        "description": "PXR-Block-Proxyサービス"
    },
    "servers": [
        {
            "url": "/pxr-block-proxy"
        }
    ],
    "paths": {
...
        "/ind": {
            "get": {
                "tags": [
                    "個人用プロキシ"
                ],
                "description": "GETメソッドのAPIを代理で呼び出します。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "access-token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "block",
                        "in": "query",
                        "schema": {
                            "oneOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "string"
                                }
                            ]
                        },
                        "description": "宛先Blockコード"
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "API URL。透過先APIにQueryParamが必要であればここに含め、URLエンコーディングを行う。"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "レスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "代理呼出先のレスポンス",
                                    "additionalProperties": true
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary",
                                    "additionalProperties": true
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "個人用プロキシ"
                ],
                "description": "POSTメソッドのAPIを代理で呼び出します。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "access-token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "block",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "宛先Blockコード"
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "API URL。透過先APIにQueryParamが必要であればここに含め、URLエンコーディングを行う。"
                    }
                ],
                "requestBody": {
                    "description": "透過先APIに必要なリクエストボディ",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "通知エンティティ",
                                "additionalProperties": true
                            }
                        },
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary",
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "代理呼出先のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "代理呼出先のレスポンス",
                                    "additionalProperties": true
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary",
                                    "additionalProperties": true
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "個人用プロキシ"
                ],
                "description": "PUTメソッドのAPIを代理で呼び出します。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "access-token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "block",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "宛先Blockコード"
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "API URL。透過先APIにQueryParamが必要であればここに含め、URLエンコーディングを行う。"
                    }
                ],
                "requestBody": {
                    "description": "透過先APIに必要なリクエストボディ",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "通知エンティティ",
                                "additionalProperties": true
                            }
                        },
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary",
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "代理呼出先のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "代理呼出先のレスポンス",
                                    "additionalProperties": true
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary",
                                    "additionalProperties": true
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "個人用プロキシ"
                ],
                "description": "DELETEメソッドのAPIを代理で呼び出します。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "access-token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "block",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "宛先Blockコード"
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "API URL。透過先APIにQueryParamが必要であればここに含め、URLエンコーディングを行う。"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "レスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "代理呼出先のレスポンス",
                                    "additionalProperties": true
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary",
                                    "additionalProperties": true
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}